Fix compilation issue on Win64.
R=yangguo@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9465003

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10818 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.cc
@@ -699,7 +699,7 @@ int MemoryAllocator::CodePageGuardStartOffset() {
 
 
 int MemoryAllocator::CodePageGuardSize() {
-  return OS::CommitPageSize();
+  return static_cast<int>(OS::CommitPageSize());
 }
 
 
@@ -713,7 +713,7 @@ int MemoryAllocator::CodePageAreaStartOffset() {
 int MemoryAllocator::CodePageAreaEndOffset() {
   // We are guarding code pages: the last OS page will be protected as
   // non-writable.
-  return Page::kPageSize - OS::CommitPageSize();
+  return Page::kPageSize - static_cast<int>(OS::CommitPageSize());
 }
